tests: benchmarks: xSPI performance - disable CPU load in CI PM test
Disable CPU load monitoring for CI PM test.
Load monitor conusmes power and alters the results.
