+++ tests/benchmarks/xspi_performance/CMakeLists.txt
@@ -9,4 +9,5 @@ cmake_minimum_required(VERSION 3.20.0)
 find_package(Zephyr REQUIRED HINTS $ENV{ZEPHYR_BASE})
 project(xspi_performance)
 
+target_sources_ifdef(CONFIG_CPU_LOAD app PRIVATE src/cpu_load_monitor.c)
 target_sources(app PRIVATE src/main.c)
