+++ tests/benchmarks/xspi_performance/prj.conf
@@ -2,7 +2,6 @@ CONFIG_FLASH=y
 CONFIG_FLASH_JESD216_API=y
 CONFIG_FLASH_PAGE_LAYOUT=y
 
-CONFIG_CPU_LOAD=y
 CONFIG_COUNTER=y
 CONFIG_DK_LIBRARY=y
